@@ -1,10 +1,10 @@
 name: wortise
 description: Wortise Flutter plugin
-version: 1.2.3
+version: 1.2.3+1
 homepage: https://wortise.com
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.20.0"
 
 dependencies:
